+++ site/src/components/Dialogs/DeleteWorkspaceDialog/DeleteWorkspaceDialog.tsx
