+++ site/src/pages/UserSettingsPage/SSHKeysPage/SSHKeysPage.tsx
@@ -1,6 +1,6 @@
 import { useActor } from "@xstate/react"
 import React, { useContext, useEffect } from "react"
-import { ConfirmDialog } from "../../../components/ConfirmDialog/ConfirmDialog"
+import { ConfirmDialog } from "../../../components/Dialogs/ConfirmDialog/ConfirmDialog"
 import { Section } from "../../../components/Section/Section"
 import { XServiceContext } from "../../../xServices/StateContext"
 import { SSHKeysPageView } from "./SSHKeysPageView"
