+++ site/src/pages/UsersPage/UsersPage.tsx
@@ -3,8 +3,8 @@ import { FC, ReactNode, useContext, useEffect } from "react"
 import { Helmet } from "react-helmet-async"
 import { useNavigate } from "react-router"
 import { useSearchParams } from "react-router-dom"
-import { ConfirmDialog } from "../../components/ConfirmDialog/ConfirmDialog"
-import { ResetPasswordDialog } from "../../components/ResetPasswordDialog/ResetPasswordDialog"
+import { ConfirmDialog } from "../../components/Dialogs/ConfirmDialog/ConfirmDialog"
+import { ResetPasswordDialog } from "../../components/Dialogs/ResetPasswordDialog/ResetPasswordDialog"
 import { userFilterQuery } from "../../util/filters"
 import { pageTitle } from "../../util/page"
 import { XServiceContext } from "../../xServices/StateContext"
